refactor: Add CRUD operations for albums in main.go
This commit adds functions to handle CRUD operations for albums in the main.go file. The getAlbums function responds with a list of all albums as JSON. The getAlbumByID function retrieves an album by its ID parameter. The postAlbums function adds a new album from JSON received in the request body. The putAlbums function replaces an existing album in the albums slice. The deleteAlbums function deletes an album from the albums slice. All functions are implemented using the gin-gonic framework for HTTP routing.